AP-7003: changing the missing value handling approach
We do not do filtering on the Java side anymore, as we need to keep all
the columns for X axis. Instead we do everything on the JS side.
For X column we just ignore missing values. For Y column - depends on
the chosen method.
Most of the corner cases are also resolved.