AP-16340: Don't retrieve view value if component is not executed
It otherwise caused an error in JS (because the JS view is empty)
and prevented the CEF JS view from being closed.
If a component is not executed and it's view being closed, the
default value will be returned as current view value.

AP-16340 (Closing the View (CEF) of a component after it's reset leads to the error)