Test fixed
DefaultNodeServiceTest.testSelectDataPointsInComponentView:
java.lang.IllegalArgumentException: Trying to call a selection
translation service of a node view which is not a table view.
 at org.knime.core.webui.node.view.NodeViewManager
 .createSelectionTranslationService(NodeViewManager.java:186)

Workflow needs to be executed because the node being used for the test
(NodeViewNodeFactory) returns a 'TableNodeView' if and only iff a there
is data (a table) available to the node model (otherwise it returns an
ordinary NodeView). And data point selection is only available for
'Table(Node)Views'.